Add `--ignore-embedded-text` (#47)
This PR adds the `--ignore-embedded-text` flag to the `scan` command so
that it is possible to ignore embedded text in document types that
support it (e.g. PDFs). The motivation for this feature is that some
PDFs have OCR results embedded that are low quality and should be
ignored.

The addition to `load_document` is tested via:

```
pytest -sv tests/test_end_to_end.py::test_run_with_ignore_embedded_text
```

Notably, the answer with Tesseract only is actually incorrect, because
it _seems_ Tesseract is missing entire columns.